TG-114: Changed the file structure and made some corrections
