--- src/Components/Admin/superAdmin.module.css
@@ -1,6 +1,0 @@
-.container {
-    display: flex;
-    flex-direction: column;
-    justify-content: space-between;
-    min-height: 100vh;
-}
